Implement SIMD on stable.
We use auto vectorization to enable SIMD under stable.
It's most likely better if we hand-roll the SIMD for the
available feature set.

Currently, the best SIMD performance can be achieved with
AVX2 and Rng512. It would be helpful if we analyse this
implementation in Godbolt and port the code to stable using
intrinsics.